feat(code): Add extra logic for admin unlock servers
Admin unlock servers were interactable at the start of the game due to their ServerDiscStorage's not checking for anything related to the admin lock.

Upon getting admin access, we set a hasDisk property on the parent ServerRack to trigger light colour changes, I've added a check for the server disc storage to also only be interactable if this is set.