feat(code): Transferring taskpad logic into separate script
Starting an attempt to generalise the UI updating logic of the Task Pad whilst maintaining the freedom to make a unique taskpad UI for each phase.
No all the relevant data can be stored in a scriptable object so I have opted for placing each phase UI canvas into a list on a Task Pad script, this should now just switch the Active canvas to whichever phase we have reached.
Any common elements like task list remain in a scriptable object for the time being