feat(code): Removed isWeaponPhase check (interim logic)
The logic added here checks for an explicit class type 'WeaponTaskPhase' which is fine for the task scripts as they are only going to be relevant for  that one task.
However there is some logic in managers that operate on a single level basis. For example the time controller setting a timer.
I will add a ticket to generalise this a bit more so we are looking for a 'type' of phase instead of a specific phase in a specific level