feat(code): Generic cleanable items
To stop the broom becoming interactable outside of downtime phase. I've added it as an inheritor of the ClearableInteractables, like the shirt and notes. But a new bug is introduced as you can sweep multiple times, which adds to the 'cleaned items' count and can finish the task just by sweeping repeatedly. Going to have to tweak this to count or monitor something else so you can keep enjoying some sweeping action without completing the task early